backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.1
python-dotenv>=1.0.1
pydantic>=2.7.0
langchain>=1.0.0
langchain-core>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langsmith>=0.1.0
pdfplumber==0.11.5
pandas>=2.2.2